ovn: merge policy_common.go back into policy.go
